Rotations & Enhanced Gravity
Gravity is 50% stronger while on an entity while it's falling. All sprites that are part of the tilemap can now be rotated in the editor by pressing left-control. Additionally, new sprites have been added to the 'dirt' tile type and a new tile type was also added. The new type, 'empty_dirt', will not have collision.